scam-detector-fixed (1)/scam-detector/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
tensorflow==2.16.1; python_version < "3.13"
librosa==0.10.2
numpy==1.26.4
soundfile==0.12.1
scipy==1.13.0
reportlab==4.2.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
websockets==12.0
Pillow==10.3.0
matplotlib==3.8.4
